--- a/samples/Sample/Program.cs
+++ b/samples/Sample/Program.cs
@@ -32,37 +32,16 @@ services.AddLogging(l => l.AddConsole());
 var serviceProvider = services.BuildServiceProvider();
 var logger = serviceProvider.GetRequiredService<ILogger<Program>>();
 
-var startTime = DateTimeOffset.UtcNow;
-logger.LogInformation(1, "Started at {StartTime} and 0x{Hello:X} is hex of 42", startTime, 42);
+var eventId = new EventId(1001, "Test");
 
-try
+for (int i = 0; i < 1_000; i++)
 {
-    throw new Exception("Boom!");
+    logger.Log(
+        LogLevel.Information,
+        eventId,
+        "Subscription {SubscriptionId} for entity {EntityName} handler for message {MessageId} has been successfully completed.",
+        "my-subscription-id",
+        "TestQueue",
+        1);
 }
-catch (Exception ex)
-{
-    logger.LogCritical(ex, "Unexpected critical error starting application");
-    logger.Log(LogLevel.Critical, 0, "Unexpected critical error", ex, null!);
-    // This write should not log anything
-    logger.Log<object>(LogLevel.Critical, 0, null!, null, null!);
-    logger.LogError(ex, "Unexpected error");
-    logger.LogWarning(ex, "Unexpected warning");
-}
-
-using (logger.BeginScope("Main"))
-{
-    logger.LogInformation("Waiting for user input");
-    var key = Console.Read();
-    logger.LogInformation("User pressed {@KeyInfo}", new { Key = key, KeyChar = (char)key });
-}
-
-var endTime = DateTimeOffset.UtcNow;
-logger.LogInformation(2, "Stopping at {StopTime}", endTime);
-
-logger.LogInformation("Stopping");
-
-logger.LogInformation("{Result,-10:l}{StartTime,15:l}{EndTime,15:l}{Duration,15:l}", "RESULT", "START TIME", "END TIME", "DURATION(ms)");
-logger.LogInformation("{Result,-10:l}{StartTime,15:l}{EndTime,15:l}{Duration,15:l}", "------", "----- ----", "--- ----", "------------");
-logger.LogInformation("{Result,-10:l}{StartTime,15:mm:s tt}{EndTime,15:mm:s tt}{Duration,15}", "SUCCESS", startTime, endTime, (endTime - startTime).TotalMilliseconds);
-
 serviceProvider.Dispose();