--- a/test/Serilog.Extensions.Logging.Benchmarks/LogEventBenchmark.cs
+++ b/test/Serilog.Extensions.Logging.Benchmarks/LogEventBenchmark.cs
@@ -34,6 +34,7 @@ namespace Serilog.Extensions.Logging.Benchmarks
 
         readonly IMelLogger _melLogger;
         readonly Person _bob, _alice;
+        readonly EventId _eventId = new EventId(1, "Test");
 
         public LogEventBenchmark()
         {
@@ -60,6 +61,17 @@ namespace Serilog.Extensions.Logging.Benchmarks
         {
             using (var scope = _melLogger.BeginScope("Hi {@User} from {$Me}", _bob, _alice))
                 _melLogger.LogInformation("Hi");
+        }
+
+        [Benchmark]
+        public void LogInformation_WithEventId()
+        {
+            this._melLogger.Log(
+                LogLevel.Information,
+                _eventId,
+                "Hi {@User} from {$Me}",
+                _bob,
+                _alice);
         }
     }
 }