Address #259 

The `LogEventConstructionBenchmark` results

BenchmarkDotNet v0.13.10, Windows 10 (10.0.19045.5073/22H2/2022Update)
Intel Core i5-10300H CPU 2.50GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK 8.0.100
  [Host]     : .NET 8.0.10 (8.0.1024.46610), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.10 (8.0.1024.46610), X64 RyuJIT AVX2

Before:
| Method       | Mean     | Error   | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------- |---------:|--------:|---------:|------:|--------:|-------:|----------:|------------:|
| Native       | 207.0 ns | 4.18 ns |  4.29 ns |  1.00 |    0.00 | 0.0706 |     296 B |        1.00 |
| NoId         | 296.4 ns | 3.47 ns |  3.07 ns |  1.43 |    0.04 | 0.1106 |     464 B |        1.57 |
| LowNumbered  | 381.9 ns | 5.28 ns |  4.68 ns |  1.84 |    0.05 | 0.1740 |     728 B |        2.46 |
| HighNumbered | 394.6 ns | 7.43 ns | 13.76 ns |  1.97 |    0.06 | 0.1931 |     808 B |        2.73 |

After:
| Method       | Mean     | Error   | StdDev  | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------- |---------:|--------:|--------:|------:|--------:|-------:|----------:|------------:|
| Native       | 209.7 ns | 4.00 ns | 3.75 ns |  1.00 |    0.00 | 0.0706 |     296 B |        1.00 |
| NoId         | 296.4 ns | 5.75 ns | 5.10 ns |  1.41 |    0.04 | 0.1106 |     464 B |        1.57 |
| LowNumbered  | 348.0 ns | 6.91 ns | 6.47 ns |  1.66 |    0.04 | 0.1335 |     560 B |        1.89 |
| HighNumbered | 332.4 ns | 3.61 ns | 3.20 ns |  1.59 |    0.03 | 0.1335 |     560 B |        1.89 |